feat: create API class and isolated REST controller to manage the Elasticsearch API endpoint
