+++ includes/class-elasticsearch-wp-rest-api-controller.php
@@ -1,326 +1,16 @@
 <?php
-// abort if not called via WordPress
-if ( ! defined( 'ABSPATH' ) ) {
-  exit;
-}
-
-if ( ! class_exists( 'WP_ES_FEEDER_REST_Controller' ) ) {
-  class WP_ES_FEEDER_REST_Controller extends WP_REST_Controller {
-    public $resource;
-    public $type;
-
-    public function __construct( $post_type ) {
-      $api_helper = new ES_Feeder\Admin\Helpers\API_Helper( $this->plugin );
-
-      $this->plugin   = 'wp-es-feeder';
-      $this->resource = $api_helper->get_post_type_label( $post_type, 'name' );
-      $this->type     = $post_type;
-    }
-
-    /**
-     * Check whether the given post should be indexed.
-     *
-     * @param object $post  A WordPress post object.
-     * @return boolean      Whether or not to index the given post.
-     *
-     * @since 3.0.0
-     */
-    public function shouldIndex( $post ) {
-      $api_helper = new \ES_Feeder\Admin\Helpers\API_Helper( $this->plugin );
-
-      return $api_helper->get_index_to_cdp( $post->ID );
-    }
-
-    public function register_routes() {
-      register_rest_route(
-        $this->namespace,
-        '/' . rawurlencode( $this->resource ),
-        array(
-          array(
-            'methods'             => WP_REST_Server::READABLE,
-            'callback'            => array(
-              $this,
-              'get_items',
-            ),
-            'args'                => array(
-              'per_page' => array(
-                'validate_callback' => function ( $param, $request, $key ) {
-                  return is_numeric( $param );
-                },
-              ),
-              'page'     => array(
-                'validate_callback' => function ( $param, $request, $key ) {
-                  return is_numeric( $param );
-                },
-              ),
-            ),
-            'permission_callback' => array(
-              $this,
-              'get_items_permissions_check',
-            ),
-          ),
-        )
-      );
-
-      register_rest_route(
-        $this->namespace,
-        '/' . rawurlencode( $this->resource ) . '/(?P<id>[\d]+)',
-        array(
-          array(
-            'methods'             => WP_REST_Server::READABLE,
-            'callback'            => array(
-              $this,
-              'get_item',
-            ),
-            'args'                => array(
-              'id' => array(
-                'validate_callback' => function ( $param, $request, $key ) {
-                  return is_numeric( $param );
-                },
-              ),
-            ),
-            'permission_callback' => array(
-              $this,
-              'get_item_permissions_check',
-            ),
-          ),
-        )
-      );
-    }
-
-    public function get_items( $request ) {
-      $args['post_type'] = $this->type;
-      $page              = (int) $request->get_param( 'page' );
-      $per_page          = (int) $request->get_param( 'per_page' );
-
-      if ( $per_page ) {
-        $args['posts_per_page'] = $per_page;
-      } else {
-        $args['posts_per_page'] = 25;
-      }
-
-      if ( is_numeric( $page ) ) {
-        if ( 1 === $page ) {
-          $args['offset'] = 0;
-        } elseif ( $page > 1 ) {
-          $args['offset'] = ( $page * $args['posts_per_page'] ) - $args['posts_per_page'];
-        }
-      }
-
-      $args['meta_query'] = array(
-        'relation' => 'OR',
-        array(
-          'key'     => '_iip_index_post_to_cdp_option',
-          'compare' => 'NOT EXISTS',
-        ),
-        array(
-          'key'     => '_iip_index_post_to_cdp_option',
-          'value'   => 'no',
-          'compare' => '!=',
-        ),
-      );
-
-      $posts = get_posts( $args );
-
-      if ( empty( $posts ) ) {
-        return rest_ensure_response( array() );
-      }
-
-      foreach ( $posts as $post ) {
-        $response = $this->prepare_item_for_response( $post, $request );
-        $data[]   = $this->prepare_response_for_collection( $response );
-      }
-
-      return rest_ensure_response( $data );
-    }
-
-    public function get_item( $request ) {
-      $api_helper = new \ES_Feeder\Admin\Helpers\API_Helper( $this->plugin );
-
-      $id       = (int) $request['id'];
-      $response = array();
-
-      $post = get_post( $id );
-
-      if ( empty( $post ) ) {
-        return rest_ensure_response( array() );
-      }
-
-      if ( $this->shouldIndex( $post ) ) {
-        $response = $this->prepare_item_for_response( $post, $request );
-        $data     = $response->get_data();
-
-        $site_taxonomies = $api_helper->get_site_taxonomies( $post->ID );
-
-        if ( count( $site_taxonomies ) ) {
-          $data['site_taxonomies'] = $site_taxonomies;
-        }
-
-        $categories = get_post_meta( $id, '_iip_taxonomy_terms', true ) ?: array();
-        $cat_ids    = array();
-        foreach ( $categories as $cat ) {
-          $args = explode( '<', $cat );
-          if ( ! in_array( $args[0], $cat_ids ) ) {
-            $cat_ids[] = $args[0];
-          }
-        }
-        $data['categories'] = $cat_ids;
-        $response->set_data( $data );
-      }
-
-      return $response;
-    }
-
-    public function prepare_response_for_collection( $response ) {
-      if ( ! ( $response instanceof WP_REST_Response ) ) {
-        return $response;
-      }
-
-      $data   = (array) $response->get_data();
-      $server = rest_get_server();
-
-      if ( method_exists( $server, 'get_compact_response_links' ) ) {
-        $links = call_user_func(
-             array(
-          $server,
-          'get_compact_response_links',
-			 ),
-            $response
-            );
-      } else {
-        $links = call_user_func(
-             array(
-          $server,
-          'get_response_links',
-			 ),
-            $response
-            );
-      }
-
-      if ( ! empty( $links ) ) {
-        $data['_links'] = $links;
-      }
-
-      return $data;
-    }
-
-    public function prepare_item_for_response( $post, $request ) {
-      return rest_ensure_response( $this->baseline( $post, $request ) );
-    }
-
-    public function baseline( $post, $request ) {
-      $api_helper = new \ES_Feeder\Admin\Helpers\API_Helper( $this->plugin );
-      $post_data  = array();
-
-      // If the post is an attachment return right away.
-      if ( 'attachment' === $post->post_type ) {
-        $post_data         = wp_prepare_attachment_for_js( $post->ID );
-        $post_data['site'] = $this->get_site();
-        return rest_ensure_response( $post_data );
-      }
-
-      // We are also renaming the fields to more understandable names.
-      if ( isset( $post->ID ) ) {
-        $post_data['post_id'] = (int) $post->ID;
-      }
-
-      $post_data['type'] = $this->type;
-
-      $post_data['site'] = $this->get_site();
-
-      $post_data['owner'] = $api_helper->get_owner( $post->ID );
-
-      if ( isset( $post->post_date ) ) {
-        $post_data['published'] = get_the_date( 'c', $post->ID );
-      }
-
-      if ( isset( $post->post_modified ) ) {
-        $post_data['modified'] = get_the_modified_date( 'c', $post->ID );
-      }
-
-      if ( isset( $post->post_author ) ) {
-        $post_data['author'] = $api_helper->get_author( $post->post_author );
-      }
-
-      // pre-approved
-      $opt               = get_option( $this->plugin );
-      $opt_url           = $opt['es_wpdomain'];
-      $post_data['link'] = str_replace( site_url(), $opt_url, get_permalink( $post->ID ) );
-
-      if ( isset( $post->post_title ) ) {
-        $post_data['title'] = $post->post_title;
-      }
-
-      if ( isset( $post->post_name ) ) {
-        $post_data['slug'] = $post->post_name;
-      }
-
-      if ( isset( $post->post_content ) ) {
-        $post_data['content'] = $api_helper->render_vc_shortcodes( $post );
-      }
-
-      if ( isset( $post->post_excerpt ) ) {
-        $post_data['excerpt'] = $post->post_excerpt;
-      }
-
-      $post_data['language']  = $api_helper->get_language( $post->ID );
-      $post_data['languages'] = $api_helper->get_related_translated_posts( $post->ID, $post->post_type ) ?: array();
-
-      if ( ! array_key_exists( 'tags', $post_data ) ) {
-        $post_data['tags'] = array();
-      }
-      if ( ! array_key_exists( 'categories', $post_data ) ) {
-        $post_data['categories'] = array();
-      }
-
-      $post_data['thumbnail'] = $api_helper->get_image_metadata( get_post_thumbnail_id( $post->ID ) );
-
-      if ( isset( $post->comment_count ) ) {
-        $post_data['comment_count'] = (int) $post->comment_count;
-      }
-
-      return $post_data;
-    }
-
-    public function get_items_permissions_check( $request ) {
-      return true;
-    }
-
-    public function get_item_permissions_check( $request ) {
-      return true;
-    }
-
-    public function authorization_status_code() {
-      $status = 401;
-      if ( is_user_logged_in() ) {
-        $status = 403;
-      }
-      return $status;
-    }
-
-    public function get_site() {
-      $opt  = get_option( $this->plugin );
-      $url  = $opt['es_wpdomain'];
-      $args = parse_url( $url );
-      $host = $url;
-      if ( array_key_exists( 'host', $args ) ) {
-        $host = $args['host'];
-      } else {
-        $host = str_ireplace( 'https://', '', str_ireplace( 'http://', '', $host ) );
-      }
-      return $host;
-    }
-  }
-
-}
-
 /**
  * Class WP_ES_FEEDER_Callback_Controller
  *
  * Handles the callback from the ES API when the sync of a post completes or fails.
+ *
+ * @since 2.0.0
  */
 class WP_ES_FEEDER_Callback_Controller {
 
+  /**
+   * @since 2.0.0
+   */
   public function register_routes() {
     register_rest_route(
       $this->namespace,
@@ -351,12 +41,14 @@ class WP_ES_FEEDER_Callback_Controller {
   /**
    * @param $request WP_REST_Request
    * @return array
+   *
+   * @since 2.0.0
    */
   public function processResponse( $request ) {
     global $wpdb, $feeder;
 
     $logger      = new ES_Feeder\Admin\Helpers\Log_Helper();
-    $sync_helper = new \ES_Feeder\Admin\Helpers\Sync_Helper( $this->plugin );
+    $sync_helper = new ES_Feeder\Admin\Helpers\Sync_Helper( $this->plugin );
     $statuses    = $sync_helper->statuses;
 
     $data = $request->get_json_params();
@@ -396,7 +88,7 @@ class WP_ES_FEEDER_Callback_Controller {
             $logger->log( "Resyncing post: $post_id, resync #$resyncs", 'callback.log' );
             update_post_meta( $post_id, '_cdp_resync_count', $resyncs );
             $post = get_post( $post_id );
-            if ( $post->post_status === 'publish' ) {
+            if ( 'publish' === $post->post_status ) {
               $feeder->post_sync_send( $post, false );
             } else {
               $feeder->delete( $post );
@@ -471,57 +163,20 @@ class WP_ES_FEEDER_Callback_Controller {
     return array( 'status' => 'ok' );
   }
 
+  /**
+   * @since 2.0.0
+   */
   public function get_items_permissions_check( $request ) {
     return true;
   }
 
+  /**
+   * @since 2.0.0
+   */
   public function get_item_permissions_check( $request ) {
     return true;
   }
 }
-
-/*
-* Creates API endpoints for all public post-types. If you have a custom post-type, you must follow
-* the class convention "WP_ES_FEEDER_EXT_{TYPE}_Controller" if you want to customize the output
-* If no class convention is found, plugin will create default API routes for custom post types
-*/
-function register_post_types( $type ) {
-  $base_types = array(
-    'post'       => true,
-    'page'       => true,
-    'attachment' => true,
-  );
-
-  $is_base_type = array_key_exists( $type, $base_types );
-  if ( (int) $is_base_type ) {
-    $controller = new WP_ES_FEEDER_REST_Controller( $type );
-    $controller->register_routes();
-    return;
-  } elseif ( ! $is_base_type && ! class_exists( 'WP_ES_FEEDER_EXT_' . strtoupper( $type ) . '_Controller' ) ) {
-    $controller = new WP_ES_FEEDER_REST_Controller( $type );
-    $controller->register_routes();
-    return;
-  }
-}
-
-function register_elasticsearch_rest_routes() {
-  $post_types = get_post_types(
-    array(
-      'public' => true,
-    )
-  );
-
-  if ( is_array( $post_types ) && count( $post_types ) > 0 ) {
-    foreach ( $post_types as $type ) {
-      register_post_types( $type );
-    }
-  }
-
-  $controller = new WP_ES_FEEDER_Callback_Controller();
-  $controller->register_routes();
-}
-
-add_action( 'rest_api_init', 'register_elasticsearch_rest_routes' );
 
 // Add cdp-rest support for the base post type.
 add_post_type_support( 'post', 'cdp-rest' );
